mock websockets in test of runner
