+++ tests/unit/test_client/test_screen_manager.py
@@ -8,6 +8,7 @@ from codejam.client.client import main_full_path
 from codejam.client.widgets.whiteboard_screen import WhiteBoardScreen
 from codejam.server.interfaces.message import Message
 from codejam.server.interfaces.topics import GameOperations
+from tests.unit.test_client.mocks import WebsocketMock
 
 
 @pytest.mark.asyncio
@@ -92,9 +93,10 @@ async def test_task_returns_ok_handler(mocker):
 
 
 @pytest.mark.asyncio
-async def test_actual_websocket_runner():
+async def test_actual_websocket_runner(mocked_websockets: WebsocketMock):
     root_widget = Builder.load_file(f"{main_full_path}")
     try:
         await asyncio.wait_for(root_widget.ids.wbs.run_websocket(), timeout=2.0)
     except asyncio.exceptions.TimeoutError:
         pass
+    assert mocked_websockets.url == f"ws://127.0.0.1:8000/ws/{root_widget.username}"
